updates json schema
