+++ util/helper.go
@@ -20,14 +20,27 @@ package util
 import (
 	"encoding/json"
 	"fmt"
-	"go.uber.org/zap"
 	"io/ioutil"
+	"math"
 	"math/rand"
 	"net/http"
+	"regexp"
 	"strconv"
 	"strings"
 	"time"
+
+	"go.uber.org/zap"
 )
+
+type resourceParser struct {
+	name        string
+	pattern     string
+	regex       *regexp.Regexp
+	conversions map[string]float64
+}
+
+var memoryParser *resourceParser
+var cpuParser *resourceParser
 
 func ContainsString(list []string, element string) bool {
 	if len(list) == 0 {
@@ -117,4 +130,100 @@ func HttpRequest(url string) (map[string]interface{}, error) {
 		return apiRes, err
 	}
 	return nil, err
+}
+
+func MemoryToNumber(memory string) (float64, error) {
+	if memoryParser == nil {
+		pattern := "(\\d*e?\\d*)(Ei?|Pi?|Ti?|Gi?|Mi?|Ki?|$)"
+		re, _ := regexp.Compile(pattern)
+		memoryParser = &resourceParser{
+			name:    "memory",
+			pattern: pattern,
+			regex:   re,
+			conversions: map[string]float64{
+				"E":  float64(1000000000000000000),
+				"P":  float64(1000000000000000),
+				"T":  float64(1000000000000),
+				"G":  float64(1000000000),
+				"M":  float64(1000000),
+				"K":  float64(1000),
+				"Ei": float64(1152921504606846976),
+				"Pi": float64(1125899906842624),
+				"Ti": float64(1099511627776),
+				"Gi": float64(1073741824),
+				"Mi": float64(1048576),
+				"Ki": float64(1024),
+			},
+		}
+	}
+	return convertResource(memoryParser, memory)
+}
+func CpuToNumber(cpu string) (float64, error) {
+	if cpuParser == nil {
+		pattern := "(\\d*e?\\d*)(m?)"
+		re, _ := regexp.Compile(pattern)
+		cpuParser = &resourceParser{
+			name:    "cpu",
+			pattern: pattern,
+			regex:   re,
+			conversions: map[string]float64{
+				"m": .001,
+			},
+		}
+	}
+	return convertResource(cpuParser, cpu)
+}
+func convertResource(rp *resourceParser, resource string) (float64, error) {
+	matches := rp.regex.FindAllStringSubmatch(resource, -1)
+	if len(matches[0]) < 2 {
+		fmt.Printf("expected pattern for %s should match %s, found %s\n", rp.name, rp.pattern, resource)
+		return float64(0), fmt.Errorf("expected pattern for %s should match %s, found %s", rp.name, rp.pattern, resource)
+	}
+	num, err := ParseFloat(matches[0][1])
+	if err != nil {
+		fmt.Println(err)
+		return float64(0), err
+	}
+	if len(matches[0]) == 3 && matches[0][2] != "" {
+		if suffix, ok := rp.conversions[matches[0][2]]; ok {
+			return num * suffix, nil
+		}
+	} else {
+		return num, nil
+	}
+	fmt.Printf("expected pattern for %s should match %s, found %s\n", rp.name, rp.pattern, resource)
+	return float64(0), fmt.Errorf("expected pattern for %s should match %s, found %s", rp.name, rp.pattern, resource)
+}
+
+func ParseFloat(str string) (float64, error) {
+	val, err := strconv.ParseFloat(str, 64)
+	if err == nil {
+		return val, nil
+	}
+
+	//Some number may be seperated by comma, for example, 23,120,123, so remove the comma firstly
+	str = strings.Replace(str, ",", "", -1)
+
+	//Some number is specifed in scientific notation
+	pos := strings.IndexAny(str, "eE")
+	if pos < 0 {
+		return strconv.ParseFloat(str, 64)
+	}
+
+	var baseVal float64
+	var expVal int64
+
+	baseStr := str[0:pos]
+	baseVal, err = strconv.ParseFloat(baseStr, 64)
+	if err != nil {
+		return 0, err
+	}
+
+	expStr := str[(pos + 1):]
+	expVal, err = strconv.ParseInt(expStr, 10, 64)
+	if err != nil {
+		return 0, err
+	}
+
+	return baseVal * math.Pow10(int(expVal)), nil
 }
